+++ crates/eww/src/app.rs
@@ -7,15 +7,18 @@ use crate::{
     paths::EwwPaths,
     script_var_handler::ScriptVarHandlerHandle,
     state::scope_graph::{ScopeGraph, ScopeIndex},
+    window_arguments::WindowArguments,
+    window_initiator::WindowInitiator,
     *,
 };
 use anyhow::anyhow;
 use codespan_reporting::files::Files;
 use eww_shared_util::{Span, VarName};
+use gdk::Monitor;
 use glib::ObjectExt;
 use itertools::Itertools;
 use once_cell::sync::Lazy;
-use simplexpr::dynval::DynVal;
+use simplexpr::{dynval::DynVal, SimplExpr};
 use std::{
     cell::RefCell,
     collections::{HashMap, HashSet},
@@ -26,7 +29,6 @@ use yuck::{
     config::{
         monitor::MonitorIdentifier,
         script_var_definition::ScriptVarDefinition,
-        window_definition::WindowDefinition,
         window_geometry::{AnchorPoint, WindowGeometry},
     },
     error::DiagError,
@@ -44,12 +46,14 @@ pub enum DaemonCommand {
     ReloadConfigAndCss(DaemonResponseSender),
     OpenInspector,
     OpenMany {
-        windows: Vec<String>,
+        windows: Vec<(String, String)>,
+        args: Vec<(String, VarName, DynVal)>,
         should_toggle: bool,
         sender: DaemonResponseSender,
     },
     OpenWindow {
         window_name: String,
+        instance_id: Option<String>,
         pos: Option<Coords>,
         size: Option<Coords>,
         anchor: Option<AnchorPoint>,
@@ -57,6 +61,7 @@ pub enum DaemonCommand {
         should_toggle: bool,
         duration: Option<std::time::Duration>,
         sender: DaemonResponseSender,
+        args: Option<Vec<(VarName, DynVal)>>,
     },
     CloseWindows {
         windows: Vec<String>,
@@ -74,12 +79,17 @@ pub enum DaemonCommand {
     },
     PrintDebug(DaemonResponseSender),
     PrintGraph(DaemonResponseSender),
-    PrintWindows(DaemonResponseSender),
+    ListWindows(DaemonResponseSender),
+    ListActiveWindows(DaemonResponseSender),
 }
 
 /// An opened window.
 #[derive(Debug)]
 pub struct EwwWindow {
+    /// Every window has an id, uniquely identifying it.
+    /// If no specific ID was specified whilst starting the window,
+    /// this will be the same as the window name.
+    pub instance_id: String,
     pub name: String,
     pub scope_index: ScopeIndex,
     pub gtk_window: gtk::Window,
@@ -104,8 +114,9 @@ pub struct App<B> {
     pub display_backend: B,
     pub scope_graph: Rc<RefCell<ScopeGraph>>,
     pub eww_config: config::EwwConfig,
-    /// Map of all currently open windows
+    /// Map of all currently open windows by their IDs
     pub open_windows: HashMap<String, EwwWindow>,
+    pub instance_id_to_args: HashMap<String, WindowArguments>,
     /// Window names that are supposed to be open, but failed.
     /// When reloading the config, these should be opened again.
     pub failed_windows: HashSet<String>,
@@ -128,6 +139,7 @@ impl<B> std::fmt::Debug for App<B> {
             .field("eww_config", &self.eww_config)
             .field("open_windows", &self.open_windows)
             .field("failed_windows", &self.failed_windows)
+            .field("window_arguments", &self.instance_id_to_args)
             .field("paths", &self.paths)
             .finish()
     }
@@ -178,14 +190,25 @@ impl<B: DisplayBackend> App<B> {
                         self.close_window(&window_name)?;
                     }
                 }
-                DaemonCommand::OpenMany { windows, should_toggle, sender } => {
+                DaemonCommand::OpenMany { windows, args, should_toggle, sender } => {
                     let errors = windows
                         .iter()
                         .map(|w| {
-                            if should_toggle && self.open_windows.contains_key(w) {
-                                self.close_window(w)
+                            let (config_name, id) = w;
+                            if should_toggle && self.open_windows.contains_key(id) {
+                                self.close_window(id)
                             } else {
-                                self.open_window(w, None, None, None, None, None)
+                                log::debug!("Config: {}, id: {}", config_name, id);
+                                let window_args = args
+                                    .iter()
+                                    .filter(|(win_id, ..)| win_id.is_empty() || win_id == id)
+                                    .map(|(_, n, v)| (n.clone(), v.clone()))
+                                    .collect();
+                                self.open_window(&WindowArguments::new_from_args(
+                                    id.to_string(),
+                                    config_name.clone(),
+                                    window_args,
+                                )?)
                             }
                         })
                         .filter_map(Result::err);
@@ -193,6 +216,7 @@ impl<B: DisplayBackend> App<B> {
                 }
                 DaemonCommand::OpenWindow {
                     window_name,
+                    instance_id,
                     pos,
                     size,
                     anchor,
@@ -200,13 +224,27 @@ impl<B: DisplayBackend> App<B> {
                     should_toggle,
                     duration,
                     sender,
+                    args,
                 } => {
-                    let is_open = self.open_windows.contains_key(&window_name);
+                    let instance_id = instance_id.unwrap_or_else(|| window_name.clone());
+
+                    let is_open = self.open_windows.contains_key(&instance_id);
+
                     let result = if should_toggle && is_open {
-                        self.close_window(&window_name)
+                        self.close_window(&instance_id)
                     } else {
-                        self.open_window(&window_name, pos, size, monitor, anchor, duration)
+                        self.open_window(&WindowArguments {
+                            instance_id,
+                            window_name,
+                            pos,
+                            size,
+                            monitor,
+                            anchor,
+                            duration,
+                            args: args.unwrap_or_default().into_iter().collect(),
+                        })
                     };
+
                     sender.respond_with_result(result)?;
                 }
                 DaemonCommand::CloseWindows { windows, sender } => {
@@ -233,16 +271,12 @@ impl<B: DisplayBackend> App<B> {
                         None => sender.send_failure(format!("Variable not found \"{}\"", name))?,
                     }
                 }
-                DaemonCommand::PrintWindows(sender) => {
-                    let output = self
-                        .eww_config
-                        .get_windows()
-                        .keys()
-                        .map(|window_name| {
-                            let is_open = self.open_windows.contains_key(window_name);
-                            format!("{}{}", if is_open { "*" } else { "" }, window_name)
-                        })
-                        .join("\n");
+                DaemonCommand::ListWindows(sender) => {
+                    let output = self.eww_config.get_windows().keys().join("\n");
+                    sender.send_success(output)?
+                }
+                DaemonCommand::ListActiveWindows(sender) => {
+                    let output = self.open_windows.iter().map(|(id, window)| format!("{id}: {}", window.name)).join("\n");
                     sender.send_success(output)?
                 }
                 DaemonCommand::PrintDebug(sender) => {
@@ -304,14 +338,14 @@ impl<B: DisplayBackend> App<B> {
     }
 
     /// Close a window and do all the required cleanups in the scope_graph and script_var_handler
-    fn close_window(&mut self, window_name: &str) -> Result<()> {
-        if let Some(old_abort_send) = self.window_close_timer_abort_senders.remove(window_name) {
+    fn close_window(&mut self, instance_id: &str) -> Result<()> {
+        if let Some(old_abort_send) = self.window_close_timer_abort_senders.remove(instance_id) {
             _ = old_abort_send.send(());
         }
         let eww_window = self
             .open_windows
-            .remove(window_name)
-            .with_context(|| format!("Tried to close window named '{}', but no such window was open", window_name))?;
+            .remove(instance_id)
+            .with_context(|| format!("Tried to close window with id '{instance_id}', but no such window was open"))?;
 
         let scope_index = eww_window.scope_index;
         eww_window.close();
@@ -324,52 +358,54 @@ impl<B: DisplayBackend> App<B> {
             self.script_var_handler.stop_for_variable(unused_var.clone());
         }
 
+        self.instance_id_to_args.remove(instance_id);
+
         Ok(())
     }
 
-    fn open_window(
-        &mut self,
-        window_name: &str,
-        pos: Option<Coords>,
-        size: Option<Coords>,
-        monitor: Option<MonitorIdentifier>,
-        anchor: Option<AnchorPoint>,
-        duration: Option<std::time::Duration>,
-    ) -> Result<()> {
-        self.failed_windows.remove(window_name);
-        log::info!("Opening window {}", window_name);
+    fn open_window(&mut self, window_args: &WindowArguments) -> Result<()> {
+        let instance_id = &window_args.instance_id;
+        self.failed_windows.remove(instance_id);
+        log::info!("Opening window {} as '{}'", window_args.window_name, instance_id);
 
         // if an instance of this is already running, close it
-        // TODO make reopening optional via a --no-reopen flag?
-        if self.open_windows.contains_key(window_name) {
-            self.close_window(window_name)?;
+        if self.open_windows.contains_key(instance_id) {
+            self.close_window(instance_id)?;
         }
 
+        self.instance_id_to_args.insert(instance_id.to_string(), window_args.clone());
+
         let open_result: Result<_> = try {
-            let mut window_def = self.eww_config.get_window(window_name)?.clone();
+            let window_name: &str = &window_args.window_name;
+
+            let window_def = self.eww_config.get_window(window_name)?.clone();
             assert_eq!(window_def.name, window_name, "window definition name did not equal the called window");
-            window_def.geometry = window_def.geometry.map(|x| x.override_if_given(anchor, pos, size));
+
+            let initiator = WindowInitiator::new(&window_def, window_args)?;
 
             let root_index = self.scope_graph.borrow().root_index;
 
+            let scoped_vars_literal = initiator.get_scoped_vars().into_iter().map(|(k, v)| (k, SimplExpr::Literal(v))).collect();
+
             let window_scope = self.scope_graph.borrow_mut().register_new_scope(
-                window_name.to_string(),
+                instance_id.to_string(),
                 Some(root_index),
                 root_index,
-                HashMap::new(),
+                scoped_vars_literal,
             )?;
 
             let root_widget = crate::widgets::build_widget::build_gtk_widget(
                 &mut self.scope_graph.borrow_mut(),
                 Rc::new(self.eww_config.get_widget_definitions().clone()),
                 window_scope,
-                window_def.widget.clone(),
+                window_def.widget,
                 None,
             )?;
 
-            let monitor_geometry = get_monitor_geometry(monitor.or_else(|| window_def.monitor.clone()))?;
+            root_widget.style_context().add_class(window_name);
 
-            let mut eww_window = initialize_window::<B>(monitor_geometry, root_widget, window_def, window_scope)?;
+            let monitor = get_gdk_monitor(initiator.monitor.clone())?;
+            let mut eww_window = initialize_window::<B>(&initiator, monitor, root_widget, window_scope)?;
             eww_window.gtk_window.style_context().add_class(window_name);
 
             // initialize script var handlers for variables. As starting a scriptvar with the script_var_handler is idempodent,
@@ -383,32 +419,32 @@ impl<B: DisplayBackend> App<B> {
 
             eww_window.destroy_event_handler_id = Some(eww_window.gtk_window.connect_destroy({
                 let app_evt_sender = self.app_evt_send.clone();
-                let window_name: String = eww_window.name.to_string();
+                let instance_id = instance_id.to_string();
                 move |_| {
                     // we don't care about the actual error response from the daemon as this is mostly just a fallback.
                     // Generally, this should get disconnected before the gtk window gets destroyed.
                     // It serves as a fallback for when the window is closed manually.
                     let (response_sender, _) = daemon_response::create_pair();
-                    let command = DaemonCommand::CloseWindows { windows: vec![window_name.clone()], sender: response_sender };
+                    let command = DaemonCommand::CloseWindows { windows: vec![instance_id.clone()], sender: response_sender };
                     if let Err(err) = app_evt_sender.send(command) {
                         log::error!("Error sending close window command to daemon after gtk window destroy event: {}", err);
                     }
                 }
             }));
 
+            let duration = window_args.duration;
             if let Some(duration) = duration {
                 let app_evt_sender = self.app_evt_send.clone();
-                let window_name = window_name.to_string();
 
                 let (abort_send, abort_recv) = futures::channel::oneshot::channel();
 
                 glib::MainContext::default().spawn_local({
-                    let window_name = window_name.clone();
+                    let instance_id = instance_id.to_string();
                     async move {
                         tokio::select! {
                             _ = glib::timeout_future(duration) => {
                                 let (response_sender, mut response_recv) = daemon_response::create_pair();
-                                let command = DaemonCommand::CloseWindows { windows: vec![window_name], sender: response_sender };
+                                let command = DaemonCommand::CloseWindows { windows: vec![instance_id.clone()], sender: response_sender };
                                 if let Err(err) = app_evt_sender.send(command) {
                                     log::error!("Error sending close window command to daemon after gtk window destroy event: {}", err);
                                 }
@@ -419,17 +455,17 @@ impl<B: DisplayBackend> App<B> {
                     }
                 });
 
-                if let Some(old_abort_send) = self.window_close_timer_abort_senders.insert(window_name, abort_send) {
+                if let Some(old_abort_send) = self.window_close_timer_abort_senders.insert(instance_id.to_string(), abort_send) {
                     _ = old_abort_send.send(());
                 }
             }
 
-            self.open_windows.insert(window_name.to_string(), eww_window);
+            self.open_windows.insert(instance_id.to_string(), eww_window);
         };
 
         if let Err(err) = open_result {
-            self.failed_windows.insert(window_name.to_string());
-            Err(err).with_context(|| format!("failed to open window `{}`", window_name))
+            self.failed_windows.insert(instance_id.to_string());
+            Err(err).with_context(|| format!("failed to open window `{}`", instance_id))
         } else {
             Ok(())
         }
@@ -448,10 +484,13 @@ impl<B: DisplayBackend> App<B> {
         self.eww_config = config;
         self.scope_graph.borrow_mut().clear(self.eww_config.generate_initial_state()?);
 
-        let window_names: Vec<String> =
+        let open_window_ids: Vec<String> =
             self.open_windows.keys().cloned().chain(self.failed_windows.iter().cloned()).dedup().collect();
-        for window_name in &window_names {
-            self.open_window(window_name, None, None, None, None, None)?;
+        for instance_id in &open_window_ids {
+            let window_arguments = self.instance_id_to_args.get(instance_id).with_context(|| {
+                format!("Cannot reopen window, initial parameters were not saved correctly for {instance_id}")
+            })?;
+            self.open_window(&window_arguments.clone())?;
         }
         Ok(())
     }
@@ -480,19 +519,20 @@ impl<B: DisplayBackend> App<B> {
 }
 
 fn initialize_window<B: DisplayBackend>(
-    monitor_geometry: gdk::Rectangle,
+    window_init: &WindowInitiator,
+    monitor: Monitor,
     root_widget: gtk::Widget,
-    window_def: WindowDefinition,
     window_scope: ScopeIndex,
 ) -> Result<EwwWindow> {
-    let window = B::initialize_window(&window_def, monitor_geometry)
-        .with_context(|| format!("monitor {} is unavailable", window_def.monitor.clone().unwrap()))?;
+    let monitor_geometry = monitor.geometry();
+    let window = B::initialize_window(window_init, monitor_geometry)
+        .with_context(|| format!("monitor {} is unavailable", window_init.monitor.clone().unwrap()))?;
 
-    window.set_title(&format!("Eww - {}", window_def.name));
+    window.set_title(&format!("Eww - {}", window_init.name));
     window.set_position(gtk::WindowPosition::None);
     window.set_gravity(gdk::Gravity::Center);
 
-    if let Some(geometry) = window_def.geometry {
+    if let Some(geometry) = window_init.geometry {
         let actual_window_rect = get_window_rectangle(geometry, monitor_geometry);
         window.set_size_request(actual_window_rect.width(), actual_window_rect.height());
         window.set_default_size(actual_window_rect.width(), actual_window_rect.height());
@@ -511,21 +551,27 @@ fn initialize_window<B: DisplayBackend>(
 
     #[cfg(feature = "x11")]
     if B::IS_X11 {
-        if let Some(geometry) = window_def.geometry {
+        if let Some(geometry) = window_init.geometry {
             let _ = apply_window_position(geometry, monitor_geometry, &window);
-            if window_def.backend_options.x11.window_type != yuck::config::backend_window_options::X11WindowType::Normal {
+            if window_init.backend_options.x11.window_type != yuck::config::backend_window_options::X11WindowType::Normal {
                 window.connect_configure_event(move |window, _| {
                     let _ = apply_window_position(geometry, monitor_geometry, window);
                     false
                 });
             }
         }
-        display_backend::set_xprops(&window, monitor_geometry, &window_def)?;
+        display_backend::set_xprops(&window, monitor, window_init)?;
     }
 
     window.show_all();
 
-    Ok(EwwWindow { name: window_def.name, gtk_window: window, scope_index: window_scope, destroy_event_handler_id: None })
+    Ok(EwwWindow {
+        instance_id: window_init.id.clone(),
+        name: window_init.name.clone(),
+        gtk_window: window,
+        scope_index: window_scope,
+        destroy_event_handler_id: None,
+    })
 }
 
 /// Apply the provided window-positioning rules to the window.
@@ -555,7 +601,7 @@ fn on_screen_changed(window: &gtk::Window, _old_screen: Option<&gdk::Screen>) {
 }
 
 /// Get the monitor geometry of a given monitor, or the default if none is given
-fn get_monitor_geometry(identifier: Option<MonitorIdentifier>) -> Result<gdk::Rectangle> {
+fn get_gdk_monitor(identifier: Option<MonitorIdentifier>) -> Result<Monitor> {
     let display = gdk::Display::default().expect("could not get default display");
     let monitor = match identifier {
         Some(ident) => {
@@ -575,7 +621,7 @@ fn get_monitor_geometry(identifier: Option<MonitorIdentifier>) -> Result<gdk::Re
             .primary_monitor()
             .context("Failed to get primary monitor from GTK. Try explicitly specifying the monitor on your window.")?,
     };
-    Ok(monitor.geometry())
+    Ok(monitor)
 }
 
 /// Returns the [Monitor][gdk::Monitor] structure corresponding to the identifer.